requirements.txt
aioserial==1.3.1
pyserial==3.5
attrs==22.1.0
